Add more rest api tests. Tweak some entity annotations.

+++ CarGameXServer-ejb/src/java/org/maguz/cargamex/ejb/ManagementBean.java
@@ -6,6 +6,7 @@
 
 package org.maguz.cargamex.ejb;
 
+import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -26,20 +27,26 @@ public abstract class ManagementBean {
     
     
     protected StatusCode merge(Object entity) {
+        if (entity == null) {
+            return StatusCode.NotFound;
+        }
         try {
             em.merge(entity);
         } catch (Exception e) {
-            logger.severe("Error merge: " + e.getMessage());
+            logger.log(Level.SEVERE, "Error merge: {0}", e.getMessage());
             return StatusCode.AuthenticationFailed;
         }
         return StatusCode.OK;
     }
     
     protected StatusCode remove(Object entity) {
+        if (entity == null) {
+            return StatusCode.NotFound;
+        }
         try {
             em.remove(em.merge(entity));
         } catch (Exception e) {
-            logger.severe("Error remove: " + e.getMessage());
+            logger.log(Level.SEVERE, "Error remove: {0}", e.getMessage());
             return StatusCode.AuthenticationFailed;
         }
         return StatusCode.OK;
